@@ -10,17 +10,21 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@chakra-ui/next-js": "^2.1.5",
     "@chakra-ui/react": "^2.8.1",
     "@emotion/react": "^11.11.1",
     "@emotion/styled": "^11.11.0",
-    "firebase": "^10.5.0",
+    "@firebase/app": "^0.9.22",
+    "@firebase/auth": "^1.3.2",
+    "firebase": "^10.5.2",
     "firebase-tools": "^12.7.0",
     "framer-motion": "^10.16.4",
-    "next": "13.5.6",
+    "next": "^14.0.0",
     "react": "^18",
     "react-dom": "^18"
   },
   "devDependencies": {
+    "@types/firebase": "^3.2.1",
     "@types/node": "^20",
     "@types/react": "^18",
     "@types/react-dom": "^18",
