@@ -1,7 +1,8 @@
 // pages/_app.js
 import { ChakraProvider } from '@chakra-ui/react'
+import type { AppProps } from 'next/app'
 
-function MyApp({ Component, pageProps }) {
+function MyApp({ Component, pageProps }: AppProps) {
   return (
     <ChakraProvider>
       <Component {...pageProps} />
